Cleanup and added some doc comments

+++ voxel_render/src/consts.rs
@@ -3,5 +3,5 @@
 pub const SCREEN_WIDTH: u32 = 2560;
 pub const SCREEN_HEIGHT: u32 = 1440;
 
-pub const RENDER_RADIUS: u8 = 7;
+pub const RENDER_RADIUS: i32 = 7;
